api/backend/ingestion/promotion/requirements.txt
boto3==1.34.0
awslambdaric==2.0.8
pandas==2.2.0
geopandas==0.14.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.28
